Issue #365: Cannot rename item if it has pipe character in its name: Add unit test
